backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
celery
redis
yfinance
pandas
pandas-ta
scikit-learn
joblib
passlib[bcrypt]
python-jose[cryptography]